+++ src/checkovRunner.ts
@@ -5,6 +5,8 @@ export interface FailedCheckovCheck {
     checkName: string;
     fileLineRange: [number, number];
     resource: string;
+    guideline: string;
+    fixedDefinition?: string; 
 }
 
 interface CheckovResponse {
@@ -18,6 +20,8 @@ interface FailedCheckovCheckRaw {
     check_name: string;
     file_line_range: [number, number];
     resource: string;
+    guideline: string;
+    fixed_definition?: string;
 }
 
 interface CheckovResponseRaw {
@@ -63,7 +67,9 @@ const parseCheckovResponse = (rawResponse: CheckovResponseRaw): CheckovResponse 
                 checkId: rawCheck.check_id, 
                 checkName: rawCheck.check_name, 
                 fileLineRange: rawCheck.file_line_range,
-                resource: rawCheck.resource 
+                resource: rawCheck.resource, 
+                guideline: rawCheck.guideline,
+                fixedDefinition: rawCheck.fixed_definition
             }))
         }
     };