+++ src/diagnostics.ts
@@ -1,18 +1,28 @@
 import * as vscode from 'vscode';
 import { FailedCheckovCheck } from './checkovRunner';
 
+export interface DiagnosticReferenceCode {
+    target: vscode.Uri;
+    value: string;
+}
+
 export const applyDiagnostics = (document: vscode.TextDocument, diagnostics: vscode.DiagnosticCollection, failedCheckovChecks: FailedCheckovCheck[]): void => {
     const foundDiagnostics: vscode.Diagnostic[] = [];
     
     for (const failure of failedCheckovChecks) {
         const line = document.lineAt(failure.fileLineRange[0] - 1); // checkov results are 1-based; these lines are 0-based
         const startPos = line.range.start.translate({ characterDelta: line.firstNonWhitespaceCharacterIndex });
+        const code: DiagnosticReferenceCode = { 
+            target: vscode.Uri.parse(failure.guideline),
+            value: failure.checkId
+        };
         
         foundDiagnostics.push({
-            message: `${failure.checkId}: ${failure.checkName}`, 
+            message: failure.checkName, 
             range: new vscode.Range(startPos, line.range.end),
             severity: vscode.DiagnosticSeverity.Warning,
-            source: 'Bridgecrew'
+            source: 'Checkov ',
+            code
         });
     }