+++ src/utils.ts
@@ -1,4 +1,7 @@
+import * as vscode from 'vscode';
 import { exec } from 'child_process';
+import { FailedCheckovCheck } from './checkovRunner';
+import { DiagnosticReferenceCode } from './diagnostics';
 
 type ExecOutput = [stdout: string, stderr: string];
 export const asyncExec = async (commandToExecute: string) : Promise<ExecOutput> => {
@@ -9,3 +12,7 @@ export const asyncExec = async (commandToExecute: string) : Promise<ExecOutput> 
         });
     });
 }; 
+
+export const createDiagnosticKey = (diagnostic: vscode.Diagnostic): string => 
+    `${(diagnostic.code as DiagnosticReferenceCode).value}-${diagnostic.range.start.line + 1}`;
+export const createCheckovKey = (checkovFail: FailedCheckovCheck): string => `${checkovFail.checkId}-${checkovFail.fileLineRange[0]}`;